enforcer: Watch for process exits through polling instead of process handles
Instead of each invocation waiting for the process handle to indicate that
the process has exited, just wait for the process manager to report that
the process has disappeared from the process list. The process manager
already has to poll for active processes, so the performance penalty of
polling has already been incurred.

Watched process handles have intermittently failed to trigger on process
exit, and have proven to be unreliable.